moving param mixing code to base since it's not too large
